requirements.txt
streamlit
pandas
matplotlib
openpyxl
reportlab
python-dotenv
requests
